Add basic makefile for buliding site
Add a Makefile that lets you push the site to GitHub pages. Currently
that's not terribly interesting because I haven't implemented publish
to actually build the html to be pushed to pages, but it's a first
step.

Also add pip-requirements.txt so that any py dependencies I use can be
tracked.