Update pip requirements and license for 2015
Removes the hard requirement of an old version of ghp-import and updates the
TODOs and LICENSE to bring them up to date with 2015, rather than leaving them
at their original 2013 versions.

Mostly bookkeeping update, but should make it clear that this is *not* an
abandoned repo, just one where most of the interesting stuff hasn't made it
to the point where I'm sharing it in public yet.